pyproject.toml
[project]
name = "SampleScraper"
version = "0.1.0"
description = "Web scraper for sample data aggregation."
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "cleo>=2.1.0",
    "pyexiftool==0.5.6",
    "requests>=2.32.3",
    "selenium>=4.32.0",
    "setuptools>=80.7.1",
    "undetected-chromedriver~=3.5.5",
]
